node details provider is moved into node provider
