+++ project/FrontEnd/collaborative_science_platform/lib/models/node_details_page/node_detailed.dart
@@ -1,4 +1,3 @@
-import 'package:collaborative_science_platform/helpers/date_to_string.dart';
 import 'package:collaborative_science_platform/models/node_details_page/node.dart';
 import 'package:collaborative_science_platform/models/node_details_page/proof.dart';
 import 'package:collaborative_science_platform/models/node_details_page/question.dart';
